Updated Queries
Original queries used a LAG function with an OVER Clause. Upon reflection in the Vaccinations section, I realized that the same results could be achieved with less lines by using an AVG function with an OVER Clause. The other advantage is that I can easily change the 7-day average to a 14-day or months long average if needed.